Fix CORS issue with conditional base64/URL fetch for image upload
